Make all adapters send a disconnect notification
All adapters now send a disconnect notification when they're done communicating
with the HTTP server. This notification won't be sent in case an exception is
thrown first, but will still be sent when an HTTP error was received from the
remote server.

Also fixes a trailing line feed in the Listener and a typo in one of the tests.